Utility script & tree view
In the model_root the link_parent-child script links arg1 as the
parent of arg2.

Tree view working, but code is far from tidy.

snapshots.py has several added functions. I think the origin and
ancestor_junctions functions are not that useful, nor is the
junctions list. first_common_ancestor() is going to be much more
useful. Needs tests to be written.